requirements-dev.txt
# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=5.0.0

# Linting and formatting
black>=24.0.0
ruff>=0.6.0
mypy>=1.11.0

# Pre-commit hooks
pre-commit>=3.8.0

# Security scanning
bandit>=1.7.9
safety>=3.2.0

# Type stubs
boto3-stubs[events,sqs]>=1.34.0
types-requests>=2.32.0
